fix to HQL queries used by ClinicStudyInfoTable and StudyClinicInfoTable

+++ eclipse_ws/biobank2/src/edu/ualberta/med/biobank/model/ClinicStudyInfo.java
@@ -8,7 +8,7 @@ public class ClinicStudyInfo {
 
     public String studyShortName;
 
-    public Long patients;
+    public int patients;
 
     public Long patientVisits;
